+++ R/lognormal.R
@@ -61,8 +61,8 @@ sd.lognormal <- function(x, ...) {
     exp(x$parameters$meanlog + 0.5 * x$parameters$sdlog^2)
 }
 
-#' @importFrom stats rlnorm
+#' @importFrom stats rlnorm qlnorm
 #' @exportS3Method
 sample_dist.lognormal <- function(x, n, ...) {
-  rlnorm(n, meanlog = x$parameters$meanlog, sdlog = x$parameters$sdlog)
+  sample_bounded(x, n, rng = rlnorm, cdf = plnorm, quantile = qlnorm)
 }